pre demo build

+++ Assets/Scripts/Interactions/SimpleBatterySystem/SimpleBatteryCharger.cs
@@ -15,7 +15,7 @@ public class SimpleBatteryCharger : MonoBehaviour, IInteractable, ILabel
 
     [Header("Battery Drain Parameters")]
     public int chargeRate = 5;
-    public int chargeIntervalSeconds = 1;
+    public float chargeIntervalSeconds = 1;
 
     [Header("Triggers")]
     public GameEvent playerDropHeldEvent;
